+ filter products by search form

+++ src/main/java/dbservice/service/ProductServiceImpl.java
@@ -1,9 +1,18 @@
 package dbservice.service;
 
+import com.google.gson.Gson;
+import dbservice.converter.CategoryConverter;
 import dbservice.converter.ProductConverter;
+import dbservice.dao.BasketDao;
+import dbservice.dao.CartDao;
+import dbservice.dao.OrderDao;
 import dbservice.dao.ProductDao;
+import dbservice.dto.CategoryDto;
 import dbservice.dto.ProductDto;
-import dbservice.entity.Product;
+import dbservice.entity.Category;
+import dbservice.result.Message;
+import dbservice.result.StatusResult;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,12 +23,26 @@ import java.util.List;
 public class ProductServiceImpl implements ProductService {
 
     @Autowired
-    ProductDao productDao;
+    private ProductDao productDao;
 
     @Autowired
-    ProductConverter productConverter;
+    private ProductConverter productConverter;
+
+    @Autowired
+    private CategoryConverter categoryConverter;
+
+    @Autowired
+    private CategoryService categoryService;
+
+    @Autowired
+    private BasketDao basketDao;
+
+    @Autowired
+    private CartDao cartDao;
 
     private List<ProductDto> topProducts;
+
+    private static final Logger logger = Logger.getLogger(ProductServiceImpl.class);
 
     @Override
     @Transactional
@@ -35,14 +58,15 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     @Transactional
-    public List<ProductDto> getByName(String name) {
-        return productConverter.convertToDtoList(productDao.getByName(name));
+    public ProductDto getByName(String name) {
+        return productConverter.convertToDto(productDao.getByName(name));
     }
 
 
     @Override
     @Transactional
-    public List<ProductDto> getByCategory(String category) {
+    public List<ProductDto> getByCategory(CategoryDto categoryDto) {
+        Category category = categoryConverter.convertToEntity(categoryDto);
         return productConverter.convertToDtoList(productDao.getByCategory(category));
     }
 
@@ -55,8 +79,14 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     @Transactional
-    public List<ProductDto> getByParams(String name, Long id_category, String brand, Integer minPrice, Integer maxPrice){
-        return productConverter.convertToDtoList(productDao.getByParams(name, id_category, brand, minPrice, maxPrice));
+    public List<ProductDto> getByParams(Long id_category, String brand, Integer minPrice, Integer maxPrice){
+        return productConverter.convertToDtoList(productDao.getByParams(id_category, brand, minPrice, maxPrice));
+    }
+
+    @Override
+    @Transactional
+    public List<ProductDto> getBySearch(String search){
+        return productConverter.convertToDtoList(productDao.getBySearch(search));
     }
 
 
@@ -76,15 +106,31 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     @Transactional
-    public void add(ProductDto productDto) {
-        productDao.add(productConverter.convertToEntity(productDto));
-
+    public StatusResult add(ProductDto newProduct, Long categoryId) {
+        CategoryDto categoryDto = categoryService.getById(categoryId);
+        newProduct.setCategory(categoryDto);
+        newProduct.setSales(0);
+        ProductDto product = getByName(newProduct.getName());
+        if (product == null) {
+            productDao.add(productConverter.convertToEntity(newProduct));
+            logger.info(String.format(StatusResult.PRODUCT_SUCCESS_CREATE.getMessage(), newProduct.getName()));
+            return StatusResult.PRODUCT_SUCCESS_CREATE;
+        }
+        logger.info(String.format(StatusResult.PRODUCT_ALREADY_EXIST.getMessage(), newProduct.getName()));
+        return StatusResult.PRODUCT_ALREADY_EXIST;
     }
 
     @Override
     @Transactional
-    public void deleteById(long id) {
-        productDao.deleteById(id);
+    public String deleteById(long id) {
+        Gson gson = new Gson();
+        if (basketDao.getOrdersForProduct(id).size() ==0
+                && cartDao.getCartItemsForProduct(id).size() ==0){
+            productDao.deleteById(id);
+            logger.info(StatusResult.PRODUCT_SUCCESS_DELETE.getMessage());
+            return gson.toJson(new Message(StatusResult.PRODUCT_SUCCESS_DELETE));
+        }
+        return gson.toJson(new Message(StatusResult.PRODUCT_FAIL_DELETE));
     }
 
     @Override
